+++ packages/bluesky/lib/src/entities/list_view.dart
@@ -10,6 +10,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 
 // 🌎 Project imports:
 import 'actor.dart';
+import 'defaults.dart';
 import 'facet.dart';
 import 'keys/ids.g.dart';
 import 'list_viewer.dart';
@@ -63,7 +64,7 @@ class ListView with _$ListView {
     @JsonKey(name: 'creator') required Actor createdBy,
 
     /// The viewer of the list.
-    required ListViewer viewer,
+    @Default(defaultListViewer) ListViewer viewer,
 
     /// The date of the indexing of the list.
     required DateTime indexedAt,
@@ -76,7 +77,11 @@ class ListView with _$ListView {
   factory ListView.fromJson(Map<String, Object?> json) =>
       _$ListViewFromJson(json);
 
+  /// Returns true if authenticated user has muted this actor,
+  /// otherwise false.
+  bool get isMuted => viewer.isMuted;
+
   /// Returns true if authenticated user has not muted yet this actor,
   /// otherwise false.
-  bool get isNotMuted => viewer.isNotMuted;
+  bool get isNotMuted => !isMuted;
 }