+++ packages/bluesky/lib/src/entities/list_view.g.dart
@@ -31,8 +31,11 @@ _$_ListView _$$_ListViewFromJson(Map json) => $checkedCreate(
           avatar: $checkedConvert('avatar', (v) => v as String?),
           createdBy: $checkedConvert('creator',
               (v) => Actor.fromJson(Map<String, Object?>.from(v as Map))),
-          viewer: $checkedConvert('viewer',
-              (v) => ListViewer.fromJson(Map<String, Object?>.from(v as Map))),
+          viewer: $checkedConvert(
+              'viewer',
+              (v) => v == null
+                  ? defaultListViewer
+                  : ListViewer.fromJson(Map<String, Object?>.from(v as Map))),
           indexedAt:
               $checkedConvert('indexedAt', (v) => DateTime.parse(v as String)),
         );